requirements.txt
colorlog==6.7.0
hydra-colorlog==1.2.0
hydra-core==1.2.0
pandas==1.5.1
tensorboard==2.10.1
torch==1.13.0
torch-tb-profiler==0.4.0
torchaudio==0.13.0
torchvision==0.14.0
tqdm==4.64.1
Flask==2.2.2
Flask-Cors==3.0.10